chore: refactor fusion modes
